More refactoring and fixes
